requirements.txt
Flask==2.0.2
Flask-Migrate==3.1.0 
Flask-SQLAlchemy==2.5.1
PyJWT==2.3.0
marshmallow==3.14.0
bcrypt==3.2.0

# for windows
psycopg2-binary

# for linux
# psycopg2